% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.corerf.R
\name{predict.corerf}
\alias{predict.corerf}
\title{Predict method for corerf objects}
\usage{
\method{predict}{corerf}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class \code{(corerf, grow)} created by the
function \code{corerf}.}

\item{newdata}{Test data of the set of covariates. A data.frame with numeric
values and factors. If missing, the out-of-bag predictions in \code{object}
is returned.}

\item{...}{Optional arguments to be passed to other methods.}
}
\value{
An object of class \code{(corerf, predict)} which is a list with the
  following components:

  \item{predicted}{Predicted covariance matrices for test data. If
    \code{newdata} is missing, OOB predictions for training observations.}
  \item{bop}{Bag of Observations for Prediction. An \eqn{nxn} matrix of
    counts.}
  \item{n}{Sample size of the test data (\code{NA}'s are omitted). If
    \code{newdata} is missing, sample size of the training set.}
  \item{xvar.names}{A character vector of the covariate names.}
  \item{yvar.names}{A character vector of the response variable names.}
}
\description{
Obtain predicted covariance matrices using a corerf forest for training or
  new data.
}
\examples{
\donttest{
## load generated example data
data(data, package = "CoReRF")
xvar.names <- colnames(data$X)
yvar.names <- colnames(data$Y)
data1 <- data.frame(data$X, data$Y)

## define train/test split
set.seed(2345)
smp <- sample(1:nrow(data1), size = round(nrow(data1)*0.6), replace = FALSE)
traindata <- data1[smp,,drop=FALSE]
testdata <- data1[-smp, xvar.names, drop=FALSE]

## formula object
formula <- as.formula(paste(paste(yvar.names, collapse="+"), ".", sep=" ~ "))

## train corerf
corerf.obj <- corerf(formula, traindata, params.rfsrc = list(ntree = 50))

## predict without new data (OOB predictions will be returned)
pred.obj <- predict(corerf.obj)
pred.oob <- pred.obj$predicted

## predict with new test data
pred.obj2 <- predict(corerf.obj, newdata = testdata)
pred <- pred.obj2$predicted
}

}
\seealso{
\code{\link{corerf}}
  \code{\link{vimp.corerf}}
  \code{\link{print.corerf}}
}
