% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.corerf.R
\name{print.corerf}
\alias{print.corerf}
\title{Print summary output of a CoReRF analysis}
\usage{
\method{print}{corerf}(x, ...)
}
\arguments{
\item{x}{An object of class \code{(corerf, grow)}, \code{(corerf, predict)}
or \code{(corerf, significancetest)}.}

\item{...}{Optional arguments to be passed to other methods.}
}
\description{
Print summary output of a CoReRF analysis. This is the default print method
  for the package.
}
\examples{
\donttest{
## load generated example data
data(data, package = "CoReRF")
xvar.names <- colnames(data$X)
yvar.names <- colnames(data$Y)
data1 <- data.frame(data$X, data$Y)

## define train/test split
set.seed(2345)
smp <- sample(1:nrow(data1), size = round(nrow(data1)*0.6), replace = FALSE)
traindata <- data1[smp,,drop=FALSE]
testdata <- data1[-smp, xvar.names, drop=FALSE]

## formula object
formula <- as.formula(paste(paste(yvar.names, collapse="+"), ".", sep=" ~ "))

## train corerf
corerf.obj <- corerf(formula, traindata, params.rfsrc = list(ntree = 50))

## print the grow object
print(corerf.obj)

## predict with new test data
pred.obj <- predict(corerf.obj, newdata = testdata)

## print the predict object
print(pred.obj)
}
}
