% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimp.corerf.R
\name{vimp.corerf}
\alias{vimp.corerf}
\alias{vimp}
\title{Variable importance for corerf objects}
\usage{
\method{vimp}{corerf}(object, ...)
}
\arguments{
\item{object}{An object of class (corerf, grow).}

\item{...}{Optional arguments to be passed to other methods.}
}
\value{
An object of class \code{(corerf, vimp)} which is a list with the
  following component:

  \item{importance}{Variable importance measures (VIMP) for covariates.}
}
\description{
Calculates variable importance measures (VIMP) for covariates for training
  data.
}
\examples{
\donttest{
## load generated example data
data(data, package = "CoReRF")
xvar.names <- colnames(data$X)
yvar.names <- colnames(data$Y)
data1 <- data.frame(data$X, data$Y)

## define train/test split
set.seed(2345)
smp <- sample(1:nrow(data1), size = round(nrow(data1)*0.6), replace = FALSE)
traindata <- data1[smp,,drop=FALSE]
testdata <- data1[-smp, xvar.names, drop=FALSE]

## formula object
formula <- as.formula(paste(paste(yvar.names, collapse="+"), ".", sep=" ~ "))

## train corerf
corerf.obj <- corerf(formula, traindata, params.rfsrc = list(ntree = 50),
  importance = TRUE)

## get the variable importance measures
vimp <- corerf.obj$importance
vimp2 <- vimp(corerf.obj)$importance
}
}
\seealso{
\code{\link{plot.vimp.corerf}}
}
